backend/requirements.txt

fastapi
uvicorn
celery
redis
requests
praw
google-api-python-client
requests_html
beautifulsoup4
newspaper3k
wikipedia-api
lxml-html-clean
snscrape
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv
nest-asyncio
openai
pymongo
python-jose
passlib
motor
bcrypt
argon2-cffi==23.1.0
nltk==3.8.1
wordcloud==1.9.3
matplotlib==3.8.0
wordcloud
fastapi-mail
playwright
aiohttp